[skip ci] [skip ci] Removing assertions
