+++ package.json
@@ -35,7 +35,7 @@
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.1",
     "@eslint/js": "^10.0.1",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "@mui/material": "^9.0.1",
     "@mui/system": "^9.0.1",
     "@mui/x-data-grid": "^9.3.0",
