+++ packages/core/package.json
@@ -165,7 +165,7 @@
     "@gmod/bgzf-filehandle": "^6.0.26",
     "@gmod/http-range-fetcher": "^5.0.7",
     "@jbrowse/jexl": "^3.0.2",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "@jbrowse/quick-lru": "^7.3.5",
     "@leeoniya/ufuzzy": "^1.0.19",
     "@mui/icons-material": "^9.0.1",
