+++ packages/core/src/gpu/GpuLifecycleMixin.ts
@@ -46,7 +46,6 @@ export function GpuLifecycleMixin() {
     .volatile(() => ({
       canvasDrawn: false,
 
-      // eslint-disable-next-line @typescript-eslint/no-unnecessary-type-assertion
       currentBackend: undefined as unknown,
       renderTick: 0,
       autorunsInstalled: false,
