+++ packages/product-core/package.json
@@ -34,7 +34,7 @@
   },
   "dependencies": {
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "@mui/material": "^9.0.1",
     "copy-to-clipboard": "^4.0.2",
     "librpc-web-mod": "^2.1.1",
