+++ packages/sv-core/package.json
@@ -34,7 +34,7 @@
   "dependencies": {
     "@gmod/vcf": "^7.0.8",
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "@jbrowse/plugin-linear-genome-view": "workspace:^",
     "@mui/material": "^9.0.1",
     "mobx": "^6.15.3",
