+++ packages/tree-sidebar/package.json
@@ -33,7 +33,7 @@
   },
   "dependencies": {
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "@jbrowse/plugin-linear-genome-view": "workspace:^",
     "@mui/material": "^9.0.1",
     "mobx": "^6.15.3",
