+++ plugins/alignments/package.json
@@ -34,7 +34,7 @@
     "@gmod/cram": "^8.2.6",
     "@jbrowse/cigar-utils": "workspace:^",
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "@jbrowse/plugin-linear-genome-view": "workspace:^",
     "@jbrowse/plugin-wiggle": "workspace:^",
     "@jbrowse/sv-core": "workspace:^",
