+++ plugins/authentication/package.json
@@ -31,7 +31,7 @@
   },
   "dependencies": {
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "@mui/material": "^9.0.1",
     "generic-filehandle2": "^2.1.9",
     "mobx": "^6.15.3"
