+++ plugins/bed/package.json
@@ -35,7 +35,7 @@
     "@gmod/bed": "^2.2.5",
     "@gmod/tabix": "^3.3.8",
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "mobx": "^6.15.3",
     "rxjs": "^7.8.2"
   },
