+++ plugins/comparative-adapters/package.json
@@ -33,7 +33,7 @@
     "@gmod/tabix": "^3.3.8",
     "@jbrowse/cigar-utils": "workspace:^",
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "@jbrowse/plugin-alignments": "workspace:^",
     "@mui/material": "^9.0.1",
     "generic-filehandle2": "^2.1.9",
