+++ plugins/data-management/package.json
@@ -32,7 +32,7 @@
   "dependencies": {
     "@gmod/ucsc-hub": "^2.0.9",
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "@jbrowse/plugin-config": "workspace:^",
     "@jbrowse/product-core": "workspace:^",
     "@mui/icons-material": "^9.0.1",
