+++ plugins/gff3/package.json
@@ -33,7 +33,7 @@
     "@flatten-js/interval-tree": "^2.0.3",
     "@gmod/tabix": "^3.3.8",
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "gff-nostream": "^3.0.9",
     "mobx": "^6.15.3",
     "rxjs": "^7.8.2"
