+++ plugins/gtf/package.json
@@ -33,7 +33,7 @@
     "@flatten-js/interval-tree": "^2.0.3",
     "@gmod/bgzf-filehandle": "^6.0.26",
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "@jbrowse/plugin-linear-genome-view": "workspace:^",
     "@mui/material": "^9.0.1",
     "gtf-nostream": "^1.3.4",
