+++ plugins/linear-genome-view/src/BaseLinearDisplay/models/fetchAutorunIntegration.test.ts
@@ -11,7 +11,6 @@ function createMockDisplayModel() {
   const state = observable({
     fetchGeneration: 0,
 
-    // eslint-disable-next-line @typescript-eslint/no-unnecessary-type-assertion
     error: undefined as unknown,
     regionTooLargeState: false,
     loadedRegions: new Map<number, Region>(),
