+++ plugins/linear-genome-view/src/BaseLinearDisplay/models/serverSideRenderedBlock.ts
@@ -102,7 +102,7 @@ const blockState = types
     /**
      * #volatile
      */
-    // eslint-disable-next-line @typescript-eslint/no-unnecessary-type-assertion
+
     error: undefined as unknown,
     /**
      * #volatile
