+++ plugins/linear-genome-view/src/LinearGenomeView/model.ts
@@ -291,12 +291,12 @@ export function stateModelFactory(pluginManager: PluginManager) {
       /**
        * #volatile
        */
-      // eslint-disable-next-line @typescript-eslint/no-unnecessary-type-assertion
+
       volatileError: undefined as unknown,
       /**
        * #volatile
        */
-      // eslint-disable-next-line @typescript-eslint/no-unnecessary-type-assertion
+
       trackRefs: {} as Record<string, HTMLDivElement>,
       /**
        * #volatile
