+++ plugins/lollipop/package.json
@@ -31,7 +31,7 @@
   },
   "dependencies": {
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "@jbrowse/plugin-linear-genome-view": "workspace:^",
     "mobx": "^6.15.3",
     "mobx-react": "^9.2.1"
