+++ plugins/menus/package.json
@@ -31,11 +31,11 @@
   },
   "dependencies": {
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "@mui/icons-material": "^9.0.1",
     "@mui/material": "^9.0.1",
     "@mui/x-data-grid": "^9.3.0",
-    "date-fns": "^4.2.1",
+    "date-fns": "^4.3.0",
     "mobx": "^6.15.3",
     "mobx-react": "^9.2.1",
     "react-dropzone": "^15.0.0"
