+++ plugins/rdf/package.json
@@ -31,7 +31,7 @@
   },
   "dependencies": {
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "mobx": "^6.15.3",
     "rxjs": "^7.8.2",
     "string-template": "^1.0.0"
