+++ plugins/spreadsheet-view/package.json
@@ -32,7 +32,7 @@
   "dependencies": {
     "@gmod/vcf": "^7.0.8",
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "@jbrowse/plugin-linear-genome-view": "workspace:^",
     "@jbrowse/plugin-variants": "workspace:^",
     "@jbrowse/sv-core": "workspace:^",
