+++ plugins/sv-inspector/package.json
@@ -31,7 +31,7 @@
   },
   "dependencies": {
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "@jbrowse/plugin-circular-view": "workspace:^",
     "@jbrowse/plugin-spreadsheet-view": "workspace:^",
     "@jbrowse/sv-core": "workspace:^",
