+++ plugins/variants/package.json
@@ -35,7 +35,7 @@
     "@gmod/tabix": "^3.3.8",
     "@gmod/vcf": "^7.0.8",
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "@jbrowse/plugin-circular-view": "workspace:^",
     "@jbrowse/plugin-linear-genome-view": "workspace:^",
     "@jbrowse/sv-core": "workspace:^",
