+++ plugins/wiggle/package.json
@@ -33,7 +33,7 @@
     "@gmod/bbi": "^9.0.17",
     "@gmod/hclust": "3.0.14",
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "@jbrowse/plugin-data-management": "workspace:^",
     "@jbrowse/plugin-linear-genome-view": "workspace:^",
     "@mui/icons-material": "^9.0.1",
