+++ products/jbrowse-desktop/package.json
@@ -42,7 +42,7 @@
     "@gmod/faidx": "^2.0.0",
     "@jbrowse/app-core": "workspace:^",
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "@jbrowse/plugin-alignments": "workspace:^",
     "@jbrowse/plugin-arc": "workspace:^",
     "@jbrowse/plugin-authentication": "workspace:^",
@@ -78,9 +78,9 @@
     "@mui/icons-material": "^9.0.1",
     "@mui/material": "^9.0.1",
     "@mui/x-data-grid": "^9.3.0",
-    "date-fns": "^4.2.1",
+    "date-fns": "^4.3.0",
     "deepmerge": "^4.3.1",
-    "dockview": "^6.4.0",
+    "dockview": "^6.5.0",
     "electron-context-menu": "^4.1.2",
     "electron-debug": "^4.1.0",
     "electron-updater": "^6.8.3",
