+++ products/jbrowse-react-app/package.json
@@ -47,7 +47,7 @@
     "@emotion/styled": "^11.14.1",
     "@jbrowse/app-core": "workspace:^",
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "@jbrowse/plugin-alignments": "workspace:^",
     "@jbrowse/plugin-arc": "workspace:^",
     "@jbrowse/plugin-authentication": "workspace:^",
@@ -82,7 +82,7 @@
     "@mui/material": "^9.0.1",
     "@types/file-saver-es": "^2.0.3",
     "base64-js": "^1.5.1",
-    "dockview-core": "^6.4.0",
+    "dockview-core": "^6.5.0",
     "file-saver-es": "^2.0.5",
     "mobx": "^6.15.3",
     "mobx-react": "^9.2.1",
