+++ products/jbrowse-web/package.json
@@ -24,7 +24,7 @@
     "@fontsource/roboto": "^5.2.10",
     "@jbrowse/app-core": "workspace:^",
     "@jbrowse/core": "workspace:^",
-    "@jbrowse/mobx-state-tree": "^5.6.9",
+    "@jbrowse/mobx-state-tree": "^5.8.6",
     "@jbrowse/plugin-alignments": "workspace:^",
     "@jbrowse/plugin-arc": "workspace:^",
     "@jbrowse/plugin-authentication": "workspace:^",
@@ -64,8 +64,8 @@
     "base64-js": "^1.5.1",
     "body-parser": "^2.2.2",
     "copy-to-clipboard": "^4.0.2",
-    "date-fns": "^4.2.1",
-    "dockview-core": "^6.4.0",
+    "date-fns": "^4.3.0",
+    "dockview-core": "^6.5.0",
     "file-saver-es": "^2.0.5",
     "generic-filehandle2": "^2.1.9",
     "idb": "^8.0.3",
