+++ products/jbrowse-web/src/tests/ReversedRegionLabels.test.tsx
@@ -41,7 +41,11 @@ test('export svg of reversed region with gene labels', async () => {
   fireEvent.click(await findByTestId(hts('gff3tabix_genes'), ...opts))
 
   // Wait for at least one canvas block to finish rendering
-  await screen.findAllByTestId(/prerendered_canvas.*done/, {}, { timeout: 30000 })
+  await screen.findAllByTestId(
+    /prerendered_canvas.*done/,
+    {},
+    { timeout: 30000 },
+  )
 
   await exportAndVerifySvg({
     findByTestId,
